backend/requirements.txt
aiohttp==3.6.2
appdirs==1.4.3
async-timeout==3.0.1
attrs==19.3.0
black==19.10b0
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
dict2xml==1.5
Django==2.2.7
django-debug-toolbar==2.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
entrypoints==0.3
flake8==3.7.9
idna==2.8
isort==4.3.21
json2xml==3.0.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.4
mysqlclient==1.4.6
numpy==1.18.0
pandas==0.25.3
pathspec==0.7.0
pipenv==2018.11.26
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.4.4
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2019.3
regex==2019.12.20
requests==2.22.0
six==1.11.0
slackbot==0.5.6
slackclient==2.5.0
slacker==0.14.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.0
urllib3==1.25.7
uWSGI==2.0.18
virtualenv==16.7.8
virtualenv-clone==0.5.3
websocket-client==0.44.0
wrapt==1.11.2
xlrd==1.2.0
xmltodict==0.11.0
yarl==1.4.2
